+++ src/libs/info/scan_gpu.cpp
@@ -33,9 +33,20 @@
 
 namespace bscan {
 
-  GPU::GPU(const std::string& vendor, const std::string& name, const std::string& driverVersion, int64_t memory_Bytes)
-      : _vendor(vendor), _name(name), _driverVersion(driverVersion) {
-    _memory_Bytes = memory_Bytes;
+  GPU::GPU(const std::string& vendor, const std::string& name, const std::string& processor, const std::string& driverVersion, const std::string& deviceID, const std::string& status, const std::string& installedDriversPath,  int64_t architecture, int64_t type, int64_t mode, int64_t availability, int64_t maxMemory, int64_t memoryBytes)
+      : _vendor(vendor), 
+        _name(name),
+        _processor(processor),
+        _deviceID(deviceID),
+        _status(status),
+        _installedDriversPath(installedDriversPath),
+        _driverVersion(driverVersion) {
+    _architecture = architecture,
+    _type = type,
+    _mode = mode,
+    _memoryBytes = memoryBytes,
+    _maxMemory = maxMemory,
+    _availability = availability;
   }
 
   std::string& GPU::vendor() {
@@ -52,6 +63,13 @@ namespace bscan {
     return _name;
   }
 
+  std::string& GPU::processor() {
+    if (_processor.empty()) {
+      _processor = getProcessor();
+    }
+    return _processor;
+  }
+
   std::string& GPU::driverVersion() {
     if (_driverVersion.empty()) {
       _driverVersion = getDriverVersion();
@@ -59,11 +77,67 @@ namespace bscan {
     return _driverVersion;
   }
 
-  int64_t GPU::memory_Bytes() {
-    if (_memory_Bytes == -1) {
-      _memory_Bytes = getMemory_Bytes();
+  std::string& GPU::deviceID() {
+    if (_deviceID.empty()) {
+      _deviceID = getDeviceID();
     }
-    return _memory_Bytes;
+    return _deviceID;
+  }
+
+  std::string& GPU::status() {
+    if (_status.empty()) {
+      _status = getStatus();
+    }
+    return _status;
+  }
+
+  std::string& GPU::installedDriversPath() {
+    if (_installedDriversPath.empty()) {
+      _installedDriversPath = getInstalledDriversPath();
+    }
+    return _installedDriversPath;
+  }
+
+  int64_t GPU::type() {
+    if (_type == -1) {
+      _type = getType();
+    }
+    return _type;
+  }
+
+  int64_t GPU::mode() {
+    if (_mode == -1) {
+      _mode = getMode();
+    }
+    return _mode;
+  }
+
+  int64_t GPU::architecture() {
+    if (_architecture == -1) {
+      _architecture = getArchitecture();
+    }
+    return _architecture;
+  }
+
+  int64_t GPU::availability() {
+    if (_availability == -1) {
+      _availability = getAvailability();
+    }
+    return _availability;
+  }
+
+  int64_t GPU::maxMemory() {
+    if (_maxMemory == -1) {
+      _maxMemory = getMaxMemory();
+    }
+    return _maxMemory;
+  }
+
+  int64_t GPU::memoryBytes() {
+    if (_memoryBytes == -1) {
+      _memoryBytes = getMemoryBytes();
+    }
+    return _memoryBytes;
   }
 
 };